+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # graphdoc-plugin-operations Change Log
 
+## 2.1.0 - June 2021
+
+* Adds `eraseByNameRegex` to allow removing specific operations from documentation, based on `name`.
+* Adds `eraseByDescriptionRegex` to allow removing specific operations from documentation, based on `description`.
+* Updates Documentation.
+
 ## 2.0.0 - May 2021
 
 * Breaking change: Removes `extractDescription` (after using it is a redundant feature).